Add fetch weather function
