+++ MANIFEST.in
@@ -4,6 +4,7 @@ include cea/plots/naming.csv
 include cea/schemas.yml
 include cea/scripts.yml
 include cea/workflows/*.yml
+include cea/datamanagement/weather_helper/weather.geojson
 
 recursive-include cea/databases *.xlsx *.xls *.txt *.csv *.epw
 recursive-include cea/interfaces/dashboard/plots/templates *.html
