+++ cea/datamanagement/weather_helper/__init__.py
@@ -1,0 +1,1 @@
+from .weather_helper import main
